+++ app/src/main/java/woowacourse/shopping/viewmodel/ProductsViewModel.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.viewmodel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import woowacourse.shopping.data.ProductRepository
+import woowacourse.shopping.data.product.ProductRepository
 import woowacourse.shopping.model.Product
 
 class ProductsViewModel : ViewModel() {
